+++ server/delphilint-server/src/main/java/au/com/integradev/delphilint/remote/sonarqube/SonarQubeComponent.java
@@ -1,0 +1,18 @@
+package au.com.integradev.delphilint.remote.sonarqube;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class SonarQubeComponent {
+  @JsonProperty private String key;
+  @JsonProperty private String path;
+
+  public String getKey() {
+    return key;
+  }
+
+  public String getPath() {
+    return path;
+  }
+}