+++ server/delphilint-server/src/test/java/au/com/integradev/delphilint/remote/SonarServerUtilsTest.java
@@ -2,18 +2,22 @@ package au.com.integradev.delphilint.remote;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 import au.com.integradev.delphilint.analysis.DelphiIssue;
 import au.com.integradev.delphilint.analysis.DelphiLintInputFile;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import org.apache.commons.lang3.NotImplementedException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -44,12 +48,12 @@ class SonarServerUtilsTest {
   void issuesAreConverted() throws SonarHostException {
     Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), Collections.emptySet(), Collections.emptySet());
+    SonarHost host = mock(SonarHost.class);
 
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(1, issues.size());
     assertEquals("rk1", issues.get(0).getRuleKey());
@@ -64,12 +68,12 @@ class SonarServerUtilsTest {
   void localIssuesHaveNoMetadata() throws SonarHostException {
     Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), Collections.emptySet(), Collections.emptySet());
+    SonarHost host = mock(SonarHost.class);
 
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(1, issues.size());
     assertNull(issues.get(0).getMetadata());
@@ -77,9 +81,7 @@ class SonarServerUtilsTest {
 
   @Test
   void unresolvedSecurityHotspotsPopulateMetadata() throws SonarHostException {
-    Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
-
-    Set<RemoteIssue> resolvedIssues =
+    Set<RemoteIssue> unresolvedIssues =
         Set.of(
             new RemoteIssue.Builder()
                 .withRuleKey("rk1")
@@ -92,12 +94,14 @@ class SonarServerUtilsTest {
                 .withServerMetadata("user1", "creation date 1")
                 .build());
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), Collections.emptySet(), resolvedIssues);
+    SonarHost host = mock(SonarHost.class);
+    when(host.getUnresolvedIssues(any(), any())).thenReturn(unresolvedIssues);
 
+    Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(1, issues.size());
     assertEquals("user1", issues.get(0).getMetadata().getAssignee());
@@ -107,9 +111,7 @@ class SonarServerUtilsTest {
 
   @Test
   void unresolvedIssuesPopulateMetadata() throws SonarHostException {
-    Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
-
-    Set<RemoteIssue> resolvedIssues =
+    Set<RemoteIssue> unresolvedIssues =
         Set.of(
             new RemoteIssue.Builder()
                 .withRuleKey("rk1")
@@ -121,12 +123,14 @@ class SonarServerUtilsTest {
                 .withServerMetadata("user1", "creation date 1")
                 .build());
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), Collections.emptySet(), resolvedIssues);
+    SonarHost host = mock(SonarHost.class);
+    when(host.getUnresolvedIssues(any(), any())).thenReturn(unresolvedIssues);
 
+    Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(1, issues.size());
     assertEquals("user1", issues.get(0).getMetadata().getAssignee());
@@ -137,8 +141,6 @@ class SonarServerUtilsTest {
   @ParameterizedTest
   @ValueSource(strings = {"OPEN", "CONFIRMED", "REOPENED", "TO_REVIEW"})
   void unresolvedIssuesAreNotPruned(String issueStatus) throws SonarHostException {
-    Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
-
     Set<RemoteIssue> unresolvedIssues =
         Set.of(
             new RemoteIssue.Builder()
@@ -149,23 +151,20 @@ class SonarServerUtilsTest {
                 .withHashFrom(FILE_A)
                 .build());
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), Collections.emptySet(), unresolvedIssues);
+    SonarHost host = mock(SonarHost.class);
+    when(host.getUnresolvedIssues(any(), any())).thenReturn(unresolvedIssues);
 
+    Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(1, issues.size());
   }
 
   @Test
   void resolvedIdenticalIssuesArePruned() throws SonarHostException {
-    Set<Issue> rawIssues =
-        Set.of(
-            buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)),
-            buildIssue(FILE_A, "rk2", "issue 2", new TextRange(9, 0, 20, 10)));
-
     Set<RemoteIssue> resolvedIssues =
         Set.of(
             new RemoteIssue.Builder()
@@ -176,12 +175,17 @@ class SonarServerUtilsTest {
                 .withHashFrom(FILE_A)
                 .build());
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), resolvedIssues, Collections.emptySet());
+    SonarHost host = mock(SonarHost.class);
+    when(host.getResolvedIssues(any(), any())).thenReturn(resolvedIssues);
 
+    Set<Issue> rawIssues =
+        Set.of(
+            buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)),
+            buildIssue(FILE_A, "rk2", "issue 2", new TextRange(9, 0, 20, 10)));
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(1, issues.size());
     assertEquals("issue 2", issues.get(0).getMessage());
@@ -189,11 +193,6 @@ class SonarServerUtilsTest {
 
   @Test
   void resolvedHotspotsArePruned() throws SonarHostException {
-    Set<Issue> rawIssues =
-        Set.of(
-            buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)),
-            buildIssue(FILE_A, "rk2", "issue 2", new TextRange(9, 0, 20, 10)));
-
     Set<RemoteIssue> resolvedIssues =
         Set.of(
             new RemoteIssue.Builder()
@@ -215,23 +214,23 @@ class SonarServerUtilsTest {
                 .withHashFrom(FILE_A)
                 .build());
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), resolvedIssues, Collections.emptySet());
+    SonarHost host = mock(SonarHost.class);
+    when(host.getResolvedIssues(any(), any())).thenReturn(resolvedIssues);
 
+    Set<Issue> rawIssues =
+        Set.of(
+            buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)),
+            buildIssue(FILE_A, "rk2", "issue 2", new TextRange(9, 0, 20, 10)));
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(0, issues.size());
   }
 
   @Test
   void acknowledgedHotspotsAreNotPruned() throws SonarHostException {
-    Set<Issue> rawIssues =
-        Set.of(
-            buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)),
-            buildIssue(FILE_A, "rk2", "issue 2", new TextRange(9, 0, 20, 10)));
-
     Set<RemoteIssue> resolvedIssues =
         Set.of(
             new RemoteIssue.Builder()
@@ -243,12 +242,17 @@ class SonarServerUtilsTest {
                 .withHashFrom(FILE_A)
                 .build());
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), resolvedIssues, Collections.emptySet());
+    SonarHost host = mock(SonarHost.class);
+    when(host.getResolvedIssues(any(), any())).thenReturn(resolvedIssues);
 
+    Set<Issue> rawIssues =
+        Set.of(
+            buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)),
+            buildIssue(FILE_A, "rk2", "issue 2", new TextRange(9, 0, 20, 10)));
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(1, issues.size());
     assertEquals("issue 2", issues.get(0).getMessage());
@@ -256,8 +260,6 @@ class SonarServerUtilsTest {
 
   @Test
   void movedResolvedIssuesArePruned() throws SonarHostException {
-    Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
-
     Set<RemoteIssue> resolvedIssues =
         Set.of(
             new RemoteIssue.Builder()
@@ -268,21 +270,20 @@ class SonarServerUtilsTest {
                 .withHash(SonarHasher.hashFileLine(FILE_A, 6))
                 .build());
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), resolvedIssues, Collections.emptySet());
+    SonarHost host = mock(SonarHost.class);
+    when(host.getResolvedIssues(any(), any())).thenReturn(resolvedIssues);
 
+    Set<Issue> rawIssues = Set.of(buildIssue(FILE_A, "rk1", "issue 1", new TextRange(6, 3, 6, 9)));
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(0, issues.size());
   }
 
   @Test
   void movedResolvedIssuesWithDifferentMessagesArePruned() throws SonarHostException {
-    Set<Issue> rawIssues =
-        Set.of(buildIssue(FILE_A, "rk1", "local issue 1", new TextRange(6, 3, 6, 9)));
-
     Set<RemoteIssue> resolvedIssues =
         Set.of(
             new RemoteIssue.Builder()
@@ -293,31 +294,32 @@ class SonarServerUtilsTest {
                 .withHash(SonarHasher.hashFileLine(FILE_A, 6))
                 .build());
 
-    SonarHost host =
-        new DummySonarHost(Collections.emptyMap(), resolvedIssues, Collections.emptySet());
+    SonarHost host = mock(SonarHost.class);
+    when(host.getResolvedIssues(any(), any())).thenReturn(resolvedIssues);
 
+    Set<Issue> rawIssues =
+        Set.of(buildIssue(FILE_A, "rk1", "local issue 1", new TextRange(6, 3, 6, 9)));
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(0, issues.size());
   }
 
   @Test
   void issuesWithNoMessageUseRuleName() throws SonarHostException {
+    SonarHost host = mock(SonarHost.class);
+    when(host.getRuleNamesByRuleKey()).thenReturn(Map.of("rk1", "Rule key 1"));
+
     Set<Issue> rawIssues =
         Set.of(
             buildIssue(FILE_A, "rk1", "", new TextRange(6, 3, 6, 9)),
             buildIssue(FILE_A, "rk1", null, new TextRange(6, 3, 6, 9)));
-
-    Map<String, String> ruleNameMap = Map.of("rk1", "Rule key 1");
-
-    SonarHost host =
-        new DummySonarHost(ruleNameMap, Collections.emptySet(), Collections.emptySet());
-
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(2, issues.size());
     assertEquals("Rule key 1", issues.get(0).getMessage());
@@ -326,79 +328,79 @@ class SonarServerUtilsTest {
 
   @Test
   void issuesWithMessageUseMessage() throws SonarHostException {
+    SonarHost host = mock(SonarHost.class);
+    when(host.getRuleNamesByRuleKey()).thenReturn(Map.of("rk1", "Rule key 1"));
+
     Set<Issue> rawIssues =
         Set.of(buildIssue(FILE_A, "rk1", "Issue message 1", new TextRange(6, 3, 6, 9)));
-
-    Map<String, String> ruleNameMap = Map.of("rk1", "Rule key 1");
-
-    SonarHost host =
-        new DummySonarHost(ruleNameMap, Collections.emptySet(), Collections.emptySet());
-
     List<DelphiIssue> issues =
         new ArrayList<>(
-            SonarServerUtils.postProcessIssues(Collections.emptySet(), rawIssues, host));
+            SonarServerUtils.postProcessIssues(
+                Collections.emptySet(), Collections.emptySet(), rawIssues, host));
 
     assertEquals(1, issues.size());
     assertEquals("Issue message 1", issues.get(0).getMessage());
   }
 
-  static class DummySonarHost implements SonarHost {
-    private final Map<String, String> _ruleNamesByRuleKey;
-    private final Collection<RemoteIssue> _resolvedIssues;
-    private final Collection<RemoteIssue> _unresolvedIssues;
+  @Test
+  void providingNullTestPathsQueriesHost() throws SonarHostException {
+    Set<String> allFilePaths = Set.of("a", "b", "c");
+    Set<String> testFilePaths = Set.of("b", "c");
+    Set<String> mainFilePaths = Set.of("a");
 
-    public DummySonarHost(
-        Map<String, String> _ruleNamesByRuleKey,
-        Collection<RemoteIssue> _resolvedIssues,
-        Collection<RemoteIssue> _unresolvedIssues) {
-      this._ruleNamesByRuleKey = _ruleNamesByRuleKey;
-      this._resolvedIssues = _resolvedIssues;
-      this._unresolvedIssues = _unresolvedIssues;
-    }
+    SonarHost host = mock(SonarHost.class);
+    when(host.getTestFilePaths()).thenReturn(testFilePaths);
 
-    @Override
-    public SonarCharacteristics getCharacteristics() {
-      throw new NotImplementedException();
-    }
+    SonarServerUtils.postProcessIssues(allFilePaths, null, Collections.emptySet(), host);
 
-    @Override
-    public Map<String, String> getRuleNamesByRuleKey() {
-      return _ruleNamesByRuleKey;
-    }
+    verify(host, times(1)).getTestFilePaths();
+    verify(host, times(1)).getResolvedIssues(mainFilePaths, testFilePaths);
+    verify(host, times(1)).getUnresolvedIssues(mainFilePaths, testFilePaths);
+  }
 
-    @Override
-    public Collection<RemoteIssue> getResolvedIssues(Collection<String> relativeFilePaths) {
-      return _resolvedIssues;
-    }
+  @Test
+  void providingNonNullTestPathsDoesNotQueryHost() throws SonarHostException {
+    Set<String> allFilePaths = Set.of("a", "b", "c");
+    Set<String> testFilePaths = Set.of("b", "c");
+    Set<String> mainFilePaths = Set.of("a");
 
-    @Override
-    public Collection<RemoteIssue> getUnresolvedIssues(Collection<String> relativeFilePaths) {
-      return _unresolvedIssues;
-    }
+    SonarHost host = mock(SonarHost.class);
+    SonarServerUtils.postProcessIssues(allFilePaths, testFilePaths, Collections.emptySet(), host);
 
-    @Override
-    public Set<RemoteRule> getRules() {
-      throw new NotImplementedException();
-    }
+    verify(host, never()).getTestFilePaths();
+    verify(host, times(1)).getResolvedIssues(mainFilePaths, testFilePaths);
+    verify(host, times(1)).getUnresolvedIssues(mainFilePaths, testFilePaths);
+  }
 
-    @Override
-    public Set<RemoteActiveRule> getActiveRules() {
-      throw new NotImplementedException();
-    }
+  @Test
+  void onlyQueriedTestFilesInProvidedPathsAreUsed() throws SonarHostException {
+    Set<String> allFilePaths = Set.of("a", "b", "c");
+    Set<String> testFilePaths = Set.of("b", "c", "d");
+    Set<String> mainFilePaths = Set.of("a");
+    Set<String> includedTestFilePaths = Set.of("b", "c");
 
-    @Override
-    public Optional<Path> getPluginJar(String pluginKey) throws SonarHostException {
-      throw new NotImplementedException();
-    }
+    SonarHost host = mock(SonarHost.class);
+    when(host.getTestFilePaths()).thenReturn(testFilePaths);
 
-    @Override
-    public Set<RemotePlugin> getDelphiPlugins() {
-      throw new NotImplementedException();
-    }
+    SonarServerUtils.postProcessIssues(allFilePaths, null, Collections.emptySet(), host);
 
-    @Override
-    public String getName() {
-      throw new NotImplementedException();
-    }
+    verify(host, times(1)).getTestFilePaths();
+    verify(host, times(1)).getResolvedIssues(mainFilePaths, includedTestFilePaths);
+    verify(host, times(1)).getUnresolvedIssues(mainFilePaths, includedTestFilePaths);
+  }
+
+  @Test
+  void onlyProvidedTestFilesInProvidedPathsAreUsed() throws SonarHostException {
+    Set<String> allFilePaths = Set.of("a", "b", "c");
+    Set<String> testFilePaths = Set.of("b", "c", "d");
+    Set<String> mainFilePaths = Set.of("a");
+    Set<String> includedTestFilePaths = Set.of("b", "c");
+
+    SonarHost host = mock(SonarHost.class);
+    SonarServerUtils.postProcessIssues(allFilePaths, testFilePaths, Collections.emptySet(), host);
+
+    verify(host, never()).getTestFilePaths();
+    verify(host, times(1)).getResolvedIssues(mainFilePaths, includedTestFilePaths);
+    verify(host, times(1)).getUnresolvedIssues(mainFilePaths, includedTestFilePaths);
   }
 }